Make sure that the runtime apis are included in metadata v15
